requirements.txt
strands-agents>=1.13.0
boto3>=1.40.0
ipython>=9.6.0
bedrock-agentcore
